src/requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
MarkupSafe==1.1.1
marshmallow==3.6.1
netaddr==0.7.20
pygal==2.4.0
PyJWT==1.7.1
python-firebase==1.2
pytz==2020.1
requests==2.24.0
six==1.15.0
timeago==1.0.14
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
